+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,166 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.3.64](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.63...@codecademy/gamut-kit@0.3.64) (2021-04-29)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.63](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.62...@codecademy/gamut-kit@0.3.63) (2021-04-29)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.62](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.61...@codecademy/gamut-kit@0.3.62) (2021-04-29)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.61](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.60...@codecademy/gamut-kit@0.3.61) (2021-04-29)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.60](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.59...@codecademy/gamut-kit@0.3.60) (2021-04-29)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.59](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.58...@codecademy/gamut-kit@0.3.59) (2021-04-29)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.58](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.57...@codecademy/gamut-kit@0.3.58) (2021-04-28)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.57](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.56...@codecademy/gamut-kit@0.3.57) (2021-04-28)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.56](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.55...@codecademy/gamut-kit@0.3.56) (2021-04-28)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.55](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.54...@codecademy/gamut-kit@0.3.55) (2021-04-27)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.54](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.53...@codecademy/gamut-kit@0.3.54) (2021-04-27)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.53](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.52...@codecademy/gamut-kit@0.3.53) (2021-04-27)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.52](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.51...@codecademy/gamut-kit@0.3.52) (2021-04-27)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.51](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.50...@codecademy/gamut-kit@0.3.51) (2021-04-26)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.50](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.49...@codecademy/gamut-kit@0.3.50) (2021-04-26)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.49](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.48...@codecademy/gamut-kit@0.3.49) (2021-04-26)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.48](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.47...@codecademy/gamut-kit@0.3.48) (2021-04-26)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.47](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.46...@codecademy/gamut-kit@0.3.47) (2021-04-26)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.46](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.45...@codecademy/gamut-kit@0.3.46) (2021-04-26)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.3.45](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.44...@codecademy/gamut-kit@0.3.45) (2021-04-23)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.3.44](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.3.43...@codecademy/gamut-kit@0.3.44) (2021-04-23)
 
 **Note:** Version bump only for package @codecademy/gamut-kit